Added messaging infrastructure using remctl.
This is done awkwardly because python-remctl is a python 2 library,
and as far as I can make out, there exists no python3-remctl (I would
love to be wrong though...) and because it's all hardcoded for now.

But the open_door function in swipedlib now takes an optional message
parameter. If non-empty, the message parameter is sent to a helper
script in Python 2 that invokes python remctl and does the right
thing.

--Ben Rosser (impersonating localadmin@janus.acm.jhu.edu)